Make libdivecomputer.c less gtk dependent
it still requires display-gtk.h for the declaration of the progressbar
type
