+++ Makefile
@@ -28,14 +28,10 @@ main.o: main.c dive.h display.h divelist.h
 	$(CC) $(CFLAGS) `pkg-config --cflags glib-2.0 gconf-2.0` \
 		-c main.c
 
-gtk-gui.o: gtk-gui.c dive.h display.h divelist.h
-	$(CC) $(CFLAGS) `pkg-config --cflags gtk+-2.0 glib-2.0 gconf-2.0` \
-		-c gtk-gui.c
-
 profile.o: profile.c dive.h display.h divelist.h
 	$(CC) $(CFLAGS) `pkg-config --cflags gtk+-2.0 glib-2.0` -c profile.c
 
-info.o: info.c dive.h display.h divelist.h
+info.o: info.c dive.h display.h display-gtk.h divelist.h
 	$(CC) $(CFLAGS) `pkg-config --cflags gtk+-2.0 glib-2.0` -c info.c
 
 equipment.o: equipment.c dive.h display.h divelist.h
@@ -44,13 +40,18 @@ equipment.o: equipment.c dive.h display.h divelist.h
 divelist.o: divelist.c dive.h display.h divelist.h
 	$(CC) $(CFLAGS) `pkg-config --cflags gtk+-2.0 glib-2.0` -c divelist.c
 
-print.o: print.c dive.h display.h
+print.o: print.c dive.h display.h display-gtk.h
 	$(CC) $(CFLAGS) `pkg-config --cflags gtk+-2.0 glib-2.0` -c print.c
 
-libdivecomputer.o: libdivecomputer.c dive.h display.h
+libdivecomputer.o: libdivecomputer.c dive.h display.h display-gtk.h libdivecomputer.h
 	$(CC) $(CFLAGS) `pkg-config --cflags gtk+-2.0 glib-2.0` \
 			-I$(LIBDIVECOMPUTERINCLUDES) \
 			-c libdivecomputer.c
+
+gtk-gui.o: gtk-gui.c dive.h display.h divelist.h display-gtk.h libdivecomputer.h
+	$(CC) $(CFLAGS) `pkg-config --cflags gtk+-2.0 glib-2.0 gconf-2.0` \
+			-I$(LIBDIVECOMPUTERINCLUDES) \
+			-c gtk-gui.c
 
 uemis.o: uemis.c uemis.h
 	$(CC) $(CFLAGS) `pkg-config --cflags gtk+-2.0 glib-2.0` -c uemis.c
